+++ src/mpl_wrap/__init__.py
@@ -3,5 +3,6 @@ from importlib import metadata
 
 __version__ = metadata.version(__name__)
 
-from mpl_wrap.wrap import *
+from mpl_wrap.data import *
+from mpl_wrap.plot import *
 from mpl_wrap.axes_wrap import *